src/components/model_training/requirements.txt
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.3.1
lightgbm==4.3.0
fastparquet==2024.5.0
joblib==1.4.2
optuna==4.2.1
pyyaml==6.0.1
